Se agrega funcionalidad sticky header + adjustments
