+++ src/scss/layout/_header.scss
@@ -27,9 +27,8 @@
   &__sticky {
     position:sticky;
     top: 20px;
-    border: 1px solid #3A3A3A;
-    background-color: white;
-    opacity: 0.8;
+    border: 1px solid var(--black);
+    background-color: var(--white);
     height: 56px;
     .header__logo--small {
       @include breakpoint (lg) {
