+++ templates/page.hbs
@@ -17,6 +17,7 @@
     {{{module header}}}
 
     <main>
+      </div>
       {{#each components}}
         {{{module this}}}
       {{/each}}
